add commonjs support
shamelesly borrowed the condition from jquery code
https://github.com/jquery/jquery/blob/8ae477a432f0924cd4bd3bdeaef2c4c15e483a8f/src/wrapper.js#L15
