+++ package.json
@@ -5,8 +5,11 @@
   "homepage": "https://github.com/san650/ceibo#readme",
   "bugs": "https://github.com/san650/ceibo/issues",
   "main": "index.js",
+  "type": "commonjs",
   "scripts": {
-    "test": "testem ci"
+    "test:browser": "testem ci",
+    "test:node": "node_modules/.bin/qunit ./test/commonjs-test.js",
+    "test": "npm run test:node && npm run test:browser"
   },
   "repository": {
     "type": "git",
