+++ test/index.html
@@ -11,7 +11,7 @@
   <script src="/testem.js" integrity="" ></script>
   <script src="../node_modules/qunit/qunit/qunit.js"></script>
   <script src="../index.js"></script>
-  <script src="tests.js"></script>
+  <script src="browser-tests.js"></script>
   <script>
     Testem.hookIntoTestFramework();
   </script>
